Receipt Entry: Added write to Firebase for separate users
Each user can submit multiple receipts to firebase, and pull the receipt
information back as a
ReceiptEntry.class.

TODO: implement queries to sort through receipts and then update the
main page with aggregated info instead of latest receipt info.

- Do this by using firebase commands for queries and saving (for each
user separately) the last referenced key, and the current values for the
main page - to reduce the amount of work required to pull from the
database.